+++ src/utils/blocks/__tests__/blocks_utils.spec.js
@@ -5,7 +5,7 @@ import {
 } from '../../../store/validators/__mocks__/validators';
 import { mockValidatorSets } from '../../../store/validators/__mocks__/validator_sets';
 import BlockProposerHandler from '../BlockProposerHandler';
-import BlocksTableAdapter from '../BlocksTableAdapter';
+import blocksTableAdapter from '../blocksTableAdapter';
 
 describe('utils/blocks', () => {
   const validatorConsensusAddressPrefix = 'did:com:valcons';
@@ -41,16 +41,17 @@ describe('utils/blocks', () => {
     expect(proposer).toStrictEqual(mockValidator());
   });
 
-  test("if 'BlocksTableAdapter' class returns an array of wanted blocks", () => {
+  test("if 'blocksTableAdapter.build' method return an array of wanted blocks", () => {
     const items = 100;
     const blocks = mockBlocks(items);
     const validators = mockValidators();
     const validatorSets = mockValidatorSets().validators;
-    const convertedBlocks = BlocksTableAdapter.setBlocks(blocks)
-      .setValidators(validators)
-      .setValidatorsSet(validatorSets)
-      .setValidatorConsensusPrefix(validatorConsensusAddressPrefix)
-      .get();
+    const tableRows = blocksTableAdapter.build({
+      blocks,
+      validators,
+      validatorsSet: validatorSets,
+      prefix: validatorConsensusAddressPrefix,
+    });
 
     const expectedKeys = [
       'height',
@@ -61,9 +62,9 @@ describe('utils/blocks', () => {
       'date',
     ];
 
-    expect(convertedBlocks.length).toBe(items);
-    for (const block of convertedBlocks) {
-      expect(Object.keys(block)).toStrictEqual(expectedKeys);
+    expect(tableRows.length).toBe(items);
+    for (const row of tableRows) {
+      expect(Object.keys(row)).toStrictEqual(expectedKeys);
     }
   });
 });